+++ NEWS.md
@@ -1,3 +1,9 @@
+# Changes to Version 1.0.3
+
+* velocity(), a new function for calculating current speed and direction have been added. 
+* Improved date handling in read.cmems() for daily data. 
+* satinMean() no longer depends on lubridate's functions.
+ 
 # Changes to Version 1.0.2
 
 * gshhg database's hyperlink changed from ftp:// to htpps://